build: Convert to wheel framework for build
This change moves the cardano_vending_machine.py to a Python module
structure and adds pyproject.toml in order to aid the build.

To install build and pytest on your local machine:

        pip3 install build pytest

To run all the tests and build the ``.whl`` file to ``dist/``:

        python3 -m pytest tests/
        python3 -m build

Note you may need to run some of these commands as sudo to achieve the
desired outcome.

[ Documentation: README.md ]
[ Testing: Manual build, deploy to cluster, venv, import module ]