./main.py: Remove protocol params from the script requirements
This is generated locally by cardano-node and is accessible from
Blockfrost instead.  Since we assume that a lot of people will get
started from scratch, we just include a convenience method in the main
script to retrieve protocol parameters, but leave in the option for the
CardanoCli to take in a file directly for others using the library.

[ Documentation: README.md ]
[ Testing: Manual ]