NftVendingMachine: Provide option for single-vend-max
The cardano-nft-mint-frontend package that is recommended for use with
this toolkit ensures that there are upper limits on the number of mints
that can occur in one transaction to help ensure there are no failures
with "TxTooLarge" that have to be debugged and refunded manually.

However, bad (or naive actors using a wallet directly) could still send
us too much money (thanks, but no thanks!) and crash our backend.  This
change provides a ``--single-vend-max`` CLI argument (as well as the
corresponding required parameter to the ``NftVendingMachine`` class)
that will limit the amount of NFTs that can be minted in a single
transaction.

This can also be used to ensure fairness too if you want to go that
route.  We recommend using this in all cases for production.

[ Documentation: README.md ]
[ Testing: Manual test with and without flag ]