Blockfrost: Incorporate Rate Limiting with Backoff
Note that the limitations are not incredibly specifically spelled out,
but there's a blurb around rate limits in the Blockfrost API that we
want to take into account if the minter is on Gigabit internet:
https://docs.blockfrost.io/#section/Limits

Typically, network RTT will be around 50ms anyway so it is unlikely we
will hit the 10 API calls per second.  But just in case the minter sells
out extremely quickly we don't want to have the vending machine
continually crash and have issues double counting completed transactions
because of that crash.

This will NOT GUARANTEE that all rate limiting issues are resolved.
Rather, it provides a bit of a buffer so that if we are going over the
burst threshold somehow, we sleep for 0.1 seconds and let Blockfrost try
to "catch its breath".

[ Documentation: None (internal scale) ]
[ Testing: Set BlockfrostApi constants to 1 & ensure print/sleep honored]