Mint: Remove the concept of "rebates" and directly calculate min-ada-val
The prior commit changed the rebate to being subtracted out of the
mint's gross profits rather than being supplied by the user.  Therefore,
we need to be much more cost conscious than merely sending back the user
a lump sum.

To calculate the rebate, we use the "min-ada-value" formula found in the
Cardano documentation here:
https://docs.cardano.org/native-tokens/minimum-ada-value-requirement

(Special thank you to mantis for his calculator to tell us if we were on
the right track: https://mantis.functionally.io/how-to/min-ada-value/).

Note that this is a bit conservative in that it uses a "coin size" of 0.
In future eras, this will be changed to 2, which will slightly lower
the minimum ada required per transaction.

[ Documentation: README.md to note the changed API/removed params ]
[ Testing: Manual QA with 1 and 6 token mints at a time ]