Whitelist: Provide initial support for an asset-based whitelist
This is a simple asset-based whitelist implementation (which could
ultimately be extended to addresses if desired).  There are three new
types introduced:

* NoWhitelist -> Performs no whitelist checks
* SingleUseWhitelist -> Allows 1 mint per whitelisted asset
* UnlimitedWhitelist -> Allows unlimited mints per whitelisted asset

The first type is simply a no-op so that we don't have "None" checks in
the main vending machine class (Mint now has a requirement that you must
pass a `Whitelist` object [duck-typed and conforming to NoWhitelist
spec]).

The second type is a directory of empty files where the name represents
the whitelisted asset identifiers (hex strings).  When a mint is
requested the whitelist returns the maximum allowed based on the UTxOs
in the transaction and the current asset ID availability.  Note that
assets are randomly selected from the input UTxO so there is no
guarantee on which are consumed if fewer mints are requested than
allowed by the whitelist.  At the end of a successful mint, the
whitelist files are moved to staging based on how many were successfully
minted.

The final type is a directory of empty files where the name represents
the whitelisted asset identifiers (hex strings), *BUT* the
implementation does not do any whitelist updating.  This corresponds to
an "unlimited whitelist" where the user must send at least one
whitelisted asset to themselves but can request as many mints as they
desire.

[ Documentation: Updated README.md with new commands ]
[ Testing: Integration testing for whitelists and expected outcomes ]