Vending Machine: Change ordering of exclusions addition to the vend loop
This is an extremely subtle change but an important one.  Consider the
following situation:

* New user UTxO is found
* Mint transaction with payouts is constructed correctly
* Transaction submitted to blockfrost but the mempool is full

In this case, a UTxO will be "stuck" in the vending machine until it is
restarted.  However, we want to discourage restart behaviors as this is
not intended to be regularly restarted mid-mint.  Instead, we move the
exclusions to *AFTER* a successful mint.  Note that this may cause an
issue whereby the metadata for the mint is marked as "used" but not
actually used, which will cause an "under-mint" of supply.

How can we be sure this won't cause a double mint? Because of the UTXO
model.  When transactions are constructed we are directly using UTXO_IN
which means we can never double spend.

WARNING: If the blockfrost API returns an error code (e.g., due to
network timeout) but the request was successfully fulfilled this may
cause us to repeatedly attempt to mint that UTxO.  No double mints will
occur, but we will start rapidly "consuming" metadata that never was
actually used.  This will cause a severe undermint.

[ Documentation: None ]
[ Testing: System integration tests at scale ]