Testing: Add initial set of tests for metadata validation
See the test cases for detailed info.  Note that we update the README to
provide information on how to run the tests under `coverage`.

There were bugs in the mint validation from commit 4d978e8 that the unit
testing revealed, so following TDD paradigms we fix it.

Currently we are at 22% code coverage.

[ Documentation: None ]
[ Testing: python3 -m coverage run --source=src/ --branch -m pytest -vv ]