Validation: Add vending machine validation for donation and rebates
There can be odd edge cases if the mint price does not exceed the
maximum possible rebate + the donation.  So, we add some pre-launch
validation to ensure that no such issue occurs and the corresponding.

In addition, we no longer support `--single-vend-max` as an optional
argument as that is impractical.  Even if the max possible is 100, we
still want users on the hook for specifying that.

Finally, we add some basic validation to ensure we are not lower than
minUTxO.

[ Documentation: Updated README.md ]
[ Testing: python3 -m coverage run --source=src/ --branch -m pytest -vv -s ]