Testing: Initial smoke tests for whitelist functionality
This essentially tests all the "plumbing" for class instantiation,
script usage, etc., without actually using the asset-based whitelist in
the test.  For now, the expected behavior is that if the UTxO did not
have the asset and the user did not send it to themselves, then no
assets are minted.

NOTE: Coverage now up to 87%

[ Documentation: None ]
[ Testing: python3 -m coverage run --source=src/ --branch -m pytest -vv -s ]