NftVendingMachine: Reverse change df04a882
Rather than potentially resubmit a transaction due to a full mempool
that is good (rare case), we take the safe approach and automatically
add new `mint_req` objects to the exclusions list up front.

NOTE: We keep the error sleep for an uncaught exception in case there is
some catastrophic error that we don't anticipate.

[ Documentation: None ]
[ Testing: TEST_ON_PREVIEW=True python3 -m coverage run --source=src/ --branch -m pytest -vv ]