Whitelist: Fix issue with multiple WL assets in 1 UTxO
Before, the check for consuming assets from the WL was done in between
UTxOs, not in between assets, which means that if a user submitted
multiple whitelist-eligible assets in the same UTxO but requested fewer
than that number, a ValueError would be raised.  By moving the check
inside of the asset loop and returning sooner we both ensure correctness
and speed up whitelist consumption.

[ Documentation: None ]
[ Testing: Full pytest suite ]