NFT Vending Machine Core: Consume whitelisting PRIOR to submitting mint
Before, we were only consuming the whitelist after the user's mint was
successfully submitted.  However, during the crush of user interactions
during the opening of a mint, we may see Blockfrost's memPool fill up,
causing rejected transactions.  In this case, since Blockfrost
automatically retries, the mint will be successful, but the user's asset
will remain on the whitelist.  The desired behavior (especially in
something like a single-use-asset-whitelist) is to pre-emptively consume
from the whitelist and then only manually intervene if a user failed to
get their NFT back.

[ Documentation: None ]
[ Testing: Full test suite ]