Blockfrost: Increase max POST retries to 2
During the Clumsy Valley mint we saw 1 single error (502) for a bad
gateway error (could have been on the EC2 side or on the
Cloudflare/Blockfrost side).  To fix this, we up the POST retries to
exactly 2 so as not to increase the delay from mempool full known errors
but also to enable a potential retry if there is a transient gateway
issue.

[ Documentation: None ]
[ Testing: None (parameters changed in pytest suite anyway) ]