Blockfrost: Switch UTxO retrieval to list
The blockfrost API guarantees (per their docs) that the ordering of the
UTxOs will be as viewed by the blockchain.  However, set() in Python has
some weird behavior.  Consider the following:

    >>> x = set()
    >>> x.add('a')
    >>> x.add('b')
    >>> x
    {'b', 'a'}
    >>> x.add('c')
    >>> x
    {'b', 'a', 'c'}

So, we change this to a list and add a test for pagination to ensure
they are processed in the correct order.

[ Documentation: <Add wiki links, documentation not already in code> ]
[ Testing: <What automated (unit, integ) and manual testing was done?> ]