Whitelist: Add simple implementation for a wallet whitelist
This change adds a new type called WalletWhitelist that requires a user
to sign with a stake or payment key and then verifies the message (using
CIP-0008).

The verification algorithm is:
1. Retrieve transaction metadata
2. Decode the 'whitelist_proof' key in 674 (CIP-0020)
3. If signature is verified via CIP-0008, retrieve the message
4. Message represents a comma-separated list of payment addresses
5. Validate that all non-reference/non-collateral addresses in the input
   transaction were signed for in the message

Note that checking the payment addresses is crucial to avoid a replay
attack by a man in the middle with some arbitrary message.

[ Documentation: pydoc, README.md, help text ]
[ Testing: pytest for new test suite ]